fix profile android ui
