+++ app/screens/profile/profile.js
@@ -13,17 +13,18 @@ import {
   Icon,
   Avatar,
 } from 'react-native-material-ui';
-import { Divider } from 'react-native-elements';
 
 // Utils
 import realm from '../../schema';
 import User from '../../utils/user';
 
-import {FontSetting} from '../../assets/style_sheets/font_setting';
-
 // Components
 import ScrollableHeader from '../../components/scrollable_header';
-import highSchoolList from '../../data/json/high_schools';
+import provinces from '../../data/json/address/provinces.json';
+import communes from '../../data/json/address/communes.json';
+import districts from '../../data/json/address/districts.json';
+import highSchools from '../../data/json/address/highSchools.json';
+import { FontSetting } from '../../assets/style_sheets/font_setting';
 
 const PROFILE_SIZE = 120;
 
@@ -45,10 +46,9 @@ export default class Profile extends Component {
   }
 
   refreshState() {
-    let user = User.getCurrent();
-    let school = highSchoolList.find((school) => school.id == user.highSchoolId);
+    let user = realm.objects('User').filtered('uuid="' + User.getID() + '"')[0];
 
-    this.setState({user: user, schoolName: !!school && school.name || ''});
+    this.setState({ user: user });
   }
 
   _renderScrollViewContent() {
@@ -67,12 +67,17 @@ export default class Profile extends Component {
         </View>
 
         { this._renderPersonalInfo() }
-        { this._renderFamilySituation() }
       </View>
     )
   }
 
   _renderPersonalInfo() {
+    let user = this.state.user;
+    let provinceName = user.provinceCode ? provinces.find((province) => province.code == user.provinceCode).label : '';
+    let districtName = user.districtCode ? districts.find((district) => district.code == user.districtCode).label : '';
+    let communeName = user.communeCode ? communes.find((commune) => commune.code == user.communeCode).label : '';
+    let schoolName = user.highSchoolCode ? highSchools.find((school) => school.code == user.highSchoolCode).label : '';
+
     return (
       <View style={[styles.box, {marginTop: 60}]}>
         <View style={styles.item}>
@@ -81,16 +86,10 @@ export default class Profile extends Component {
             <Icon name="edit" />
           </TouchableOpacity>
         </View>
-        <Divider />
 
         <View style={styles.item}>
           <Text style={styles.itemLabel}>ឈ្មោះពេញ</Text>
           <Text style={styles.itemValue}>: {this.state.user.fullName}</Text>
-        </View>
-
-        <View style={styles.item}>
-          <Text style={styles.itemLabel}>ឈ្មោះគណនី</Text>
-          <Text style={styles.itemValue}>: {this.state.user.username}</Text>
         </View>
 
         <View style={styles.item}>
@@ -104,11 +103,6 @@ export default class Profile extends Component {
         </View>
 
         <View style={styles.item}>
-          <Text style={styles.itemLabel}>សញ្ជាតិ</Text>
-          <Text style={styles.itemValue}>: {this.state.user.nationality}</Text>
-        </View>
-
-        <View style={styles.item}>
           <Text style={styles.itemLabel}>លេខទូរស័ព្ទ</Text>
           <Text style={styles.itemValue}>: {this.state.user.phoneNumber || '-'}</Text>
         </View>
@@ -119,63 +113,23 @@ export default class Profile extends Component {
         </View>
 
         <View style={styles.item}>
+          <Text style={styles.itemLabel}>ខេត្ត</Text>
+          <Text style={styles.itemValue}>: {provinceName}</Text>
+        </View>
+
+        <View style={styles.item}>
+          <Text style={styles.itemLabel}>ស្រុក</Text>
+          <Text style={styles.itemValue}>: {districtName}</Text>
+        </View>
+
+        <View style={styles.item}>
+          <Text style={styles.itemLabel}>ឃុំ</Text>
+          <Text style={styles.itemValue}>: {communeName}</Text>
+        </View>
+
+        <View style={styles.item}>
           <Text style={styles.itemLabel}>រៀននៅសាលា</Text>
-          <Text style={styles.itemValue}>: {this.state.schoolName}</Text>
-        </View>
-
-        <View style={styles.item}>
-          <Text style={styles.itemLabel}>អាស័យដ្ឋានបច្ចុប្បន្ន</Text>
-          <Text style={styles.itemValue}>: {this.state.user.address}</Text>
-        </View>
-      </View>
-    )
-  }
-
-  _renderFamilySituation() {
-    return (
-      <View style={styles.box}>
-        <View style={styles.item}>
-          <Text style={styles.itemTitle}>ស្ថានភាពគ្រួសារ</Text>
-          <TouchableOpacity onPress={() => this.props.navigation.navigate('EditFamilySituation', { refresh: this.refreshState.bind(this) })}>
-            <Icon name="edit" />
-          </TouchableOpacity>
-        </View>
-        <Divider />
-
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>ឪពុកម្តាយលែងលះ</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isDivorce ? 'លែងលះ' : 'គ្មានទេ'}</Text>
-        </View>
-
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>ពិការភាពក្នុងគ្រួសារ</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isDisable ? 'មាន' : 'គ្មានទេ'}</Text>
-        </View>
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>អំពើហិង្សាក្នុងគ្រួសារ</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isDomesticViolence ? 'មាន' : 'គ្មានទេ'}</Text>
-        </View>
-
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>សាមាជិកគ្រួសារណាមួយជក់បារី</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isSmoking ? 'មាន' : 'គ្មានទេ'}</Text>
-        </View>
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>សាមាជិកគ្រួសារណាមួយញៀនសុរា</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isAlcoholic ? 'មាន' : 'គ្មានទេ'}</Text>
-        </View>
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>សាមាជិកគ្រួសារណាមួយជក់គ្រឿងញៀន</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.isDrug ? 'មាន' : 'គ្មានទេ'}</Text>
-        </View>
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>ប្រភេទផ្ទះរបស់សិស្ស</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.houseType}</Text>
-        </View>
-
-        <View style={styles.item}>
-          <Text style={[styles.itemLabel, {flex: 2}]}>ចំណូលប្រចាំខែគិតជាលុយរៀល</Text>
-          <Text style={[styles.itemValue, {flex: 1}]}>: {this.state.user.collectiveIncome}</Text>
+          <Text style={styles.itemValue}>: {schoolName}</Text>
         </View>
       </View>
     )
@@ -228,6 +182,14 @@ const styles = StyleSheet.create({
     borderWidth: 1,
     borderColor: '#eee',
     backgroundColor: '#fff',
+    ...Platform.select({
+      android: {
+        marginHorizontal: 16,
+      },
+      ios: {
+        marginHorizontal: 0,
+      }
+    })
   },
   item: {
     flexDirection: 'row',
